requirements.txt
discord.py
ndjson
python-dateutil
python-dotenv
requests